+++ docs/ChangeLog.md
@@ -6,6 +6,10 @@
 
 * Autoloading for libraries
 
+### Removed
+
+* 4th param `$callable` of `$this->request()` and `$this->ajaxRequest()` (deprecated since pre v0.3.0)
+
 ## v0.6.2 (2015/08/13)
 
 ### Fixed
@@ -76,7 +80,7 @@
 
 ### Deprecated
 
-* 4th param `$callable` of `$this->request()` and  `$this->ajaxRequest()`. Use `$this->request->setCallable()` method instead. See [Function/Class Reference](https://github.com/kenjis/ci-phpunit-test/blob/v0.3.0/docs/FunctionAndClassReference.md#testcaserequestmethod-argv-params---callable--null).
+* 4th param `$callable` of `$this->request()` and `$this->ajaxRequest()`. Use `$this->request->setCallable()` method instead. See [Function/Class Reference](https://github.com/kenjis/ci-phpunit-test/blob/v0.3.0/docs/FunctionAndClassReference.md#testcaserequestmethod-argv-params---callable--null).
 
 ### Added